Events

+++ include/ecs/EntityImpl.hh
@@ -83,29 +83,41 @@ namespace ecs
 	template <typename CompType, typename ...T>
 	Handle<CompType> Entity::Assign(T... args)
 	{
+		if (em == nullptr) {
+			throw runtime_error("Cannot assign component to NULL Entity");
+		}
 		return em->Assign<CompType>(this->eid, args...);
 	}
 
 	template <typename CompType>
 	void Entity::Remove()
 	{
+		if (em == nullptr) {
+			throw runtime_error("Cannot remove component from NULL Entity");
+		}
 		em->Remove<CompType>(this->eid);
 	}
 
 	template <typename CompType>
 	bool Entity::Has() const
 	{
-		return em->Has<CompType>(this->eid);
+		return (em != nullptr) && em->Has<CompType>(this->eid);
 	}
 
 	template <typename CompType>
 	Handle<CompType> Entity::Get()
 	{
+		if (em == nullptr) {
+			throw runtime_error("NULL entity has no components");
+		}
 		return em->Get<CompType>(this->eid);
 	}
 
 	inline void Entity::Destroy()
 	{
+		if (em == nullptr) {
+			return;
+		}
 		em->Destroy(this->eid);
 	}
 
@@ -116,6 +128,9 @@ namespace ecs
 
 	inline void Entity::RemoveAllComponents()
 	{
+		if (em == nullptr) {
+			return;
+		}
 		em->RemoveAllComponents(this->eid);
 	}
 
@@ -128,18 +143,20 @@ namespace ecs
 	{
 		return this->eid;
 	}
-}
 
-namespace std
-{
-	// allow Entity class to be used in hashed data structures
-	template <>
-	struct hash<ecs::Entity>
+	template <typename Event>
+	Subscription Entity::Subscribe(
+		std::function<void(Entity, const Event &)> callback)
 	{
-	public:
-		size_t operator()(const ecs::Entity &e) const
-		{
-			return hash<uint64>()(e.eid.id);
+		if (em == nullptr) {
+			throw runtime_error("Cannot subscribe to events on NULL entity");
 		}
-	};
+		return em->Subscribe(callback, this->eid);
+	}
+
+	template <typename Event>
+	void Entity::Emit(const Event &event)
+	{
+		em->Emit(this->eid, event);
+	}
 }
